project0: setup env on x86_64 Ubuntu-13.04
geekos is a tiny OS for x86 pc, we should build 32bit code on 64bit
host, add FLAG "-m32" to gcc for x86 object target and "-m elf_i386" to
ld for x86 elf execution on bochs.

disable -Werror and enable -fno-stack-protector in gcc building, this
will disable some checking and optimization, but ok to build it.

bochs config file is older than working on new bochs in ubuntu now, some
adjustment employed, and this fixup bochs display in ubuntu.
http://overflowedstack.wordpress.com/2012/05/07/solve-bochs-crash-on-start-problem-under-ubuntu-12-04/

Signed-off-by: heljoy <hustliujp@gmail.com>